+++ src/biodwh2-aact/src/main/java/de/unibi/agbi/biodwh2/aact/etl/AACTGraphExporter.java
@@ -36,7 +36,7 @@ public class AACTGraphExporter extends GraphExporter<AACTDataSource> {
 
     @Override
     public long getExportVersion() {
-        return 1;
+        return 2;
     }
 
     @Override